IOS-4366 Revert 0d56cde1dc63c86759a5454f19ebdab4102ac516

+++ BlockchainSdk/WalletManagers/Ethereum/EthereumWalletManager.swift
@@ -100,7 +100,7 @@ private extension EthereumWalletManager {
                 let gasLimit = ethereumFeeResponse.gasLimit
                 let fees = ethereumFeeResponse.gasPrices.map { gasPrice in
                     let feeValue = gasLimit * gasPrice
-                    let fee = Decimal(Double(feeValue)) / self.wallet.blockchain.decimalValue
+                    let fee = Decimal(Int(feeValue)) / self.wallet.blockchain.decimalValue
 
                     let amount = Amount(with: self.wallet.blockchain, value: fee)
                     let parameters = EthereumFeeParameters(gasLimit: gasLimit, gasPrice: gasPrice)
